Abstract Estimate and Call
